requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Limiter==2.3.1
pyjwt==2.3.0